check before moving file

+++ dnsmasq/install.sh
@@ -9,8 +9,10 @@ fi
 ls res > /dev/null 2> /dev/null
 if [[ $? -eq 0 ]]; then
 	source res/check.sh
+	CONF_PATH="dnsmasq"
 else
 	source ../res/check.sh
+	CONF_PATH="."
 fi
 
 CONFIG_DIR="/etc"
@@ -21,11 +23,13 @@ sudo apt-get install dnsmasq >> $LOG_FILE 2>> $LOG_FILE
 check
 
 # backup default configuration
-echo -n "Back-up default dnsmaq configuration: "
-sudo mv $CONFIG_DIR/dnsmasq.conf $CONFIG_DIR/dnsmasq.conf.back >> $LOG_FILE 2>> $LOG_FILE
-check
+if [[ -f "$CONFIG_DIR/dnsmasq.conf" ]]; then
+	echo -n "Back-up default dnsmaq configuration: "
+	sudo mv $CONFIG_DIR/dnsmasq.conf $CONFIG_DIR/dnsmasq.conf.back >> $LOG_FILE 2>> $LOG_FILE
+	check
+fi
 
 # put our dnsmaq.conf file
 echo -n "Copying config file"
-sudo mv ./dnsmasq.conf $CONFIG_DIR >> $LOG_FILE 2>> $LOG_FILE
+sudo mv $CONF_PATH/dnsmasq.conf $CONFIG_DIR >> $LOG_FILE 2>> $LOG_FILE
 check
